+++ src/lib/components/CopyButton.tsx
@@ -36,7 +36,10 @@ export const CopyButton = ({
           variant="outline-info"
           size={size}
           float="right"
-          onClick={onCopy}
+          onClick={() => {
+            throw new Error("FRONTEND ERROR");
+            onCopy();
+          }}
           leftIcon={<CopyIcon boxSize="4" onClick={onCopy} />}
         >
           Copy